:art: Split and refactor common lambda code
